+++ NEWS.md
@@ -1,3 +1,5 @@
 # covr2mddemo (development version)
 
-* Initial CRAN submission.
+* `add_three()` adds three to a numeric vector.
+* `add_two()` adds two to a numeric vector.
+* `add_one()` adds one to a numeric vector.